log del filtro a usar

+++ daemon/server/include/Tools.h
@@ -13,7 +13,7 @@
     kernel_size:    Size of the kernel used
 
 */
-uint8_t* get_values(uint8_t* image, int* current_pos, uint8_t kernel_size){
+uint8_t* get_values(uint8_t* image, size_t* current_pos, uint8_t kernel_size){
 
     // Calc the amount of steps to move in each direction
     // Multiply by the CHANNELS to take them into account when
@@ -32,11 +32,11 @@ uint8_t* get_values(uint8_t* image, int* current_pos, uint8_t kernel_size){
     }
 
     // Get the values
-    for (int h = current_pos[1] - steps_back*WIDTH*CHANNELS; h <= current_pos[1] + steps_fwd*WIDTH*CHANNELS; h += WIDTH*CHANNELS)
+    for (size_t h = current_pos[1] - steps_back*WIDTH*CHANNELS; h <= current_pos[1] + steps_fwd*WIDTH*CHANNELS; h += WIDTH*CHANNELS)
     {
         // If we are in a valid position search for the value
         if((h >= 0) && (h < HEIGHT*WIDTH*CHANNELS)){
-            for (int w = current_pos[0] - steps_back*CHANNELS; w <= current_pos[0] + steps_fwd*CHANNELS; w += CHANNELS)
+            for (size_t w = current_pos[0] - steps_back*CHANNELS; w <= current_pos[0] + steps_fwd*CHANNELS; w += CHANNELS)
             {
                 // If we are in a valid position search for the value
                 if((w >= 0) && (w < WIDTH*CHANNELS)){
